Revert "Add UserStorageParser class, Exception Handling and JUnit Testing"
