+++ src/main/java/seedu/duke/user/UserUniversityListManager.java
@@ -2,10 +2,6 @@ package seedu.duke.user;
 
 import seedu.duke.ui.Ui;
 import seedu.duke.userstorage.UserStorage;
-import seedu.duke.exceptions.InvalidUserCommandException;
-import seedu.duke.exceptions.InvalidUserStorageFileException;
-import seedu.duke.ui.Ui;
-import seedu.duke.userstorage.UserStorageParser;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -29,13 +25,7 @@ public class UserUniversityListManager {
     }
 
     public UserUniversityListManager(String fileContent) {
-        try {
-            myManager = UserStorageParser.convertFileContentIntoUniversityList(fileContent);
-        } catch (InvalidUserStorageFileException e) {
-            System.out.println("Invalid file format in data/duke.txt");
-            System.out.println("Creating new University List Manager");
-            myManager = new HashMap<String, UserUniversityList>();
-        }
+        myManager = UserStorage.convertFileContentIntoUniversityList(fileContent);
     }
 
     /**
